# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "@yd/api",
    "summary": "API to search and download YouTube videos in various formats.",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/jordanshatford/youtube-downloader"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/jordanshatford/youtube-downloader/blob/main/LICENSE"
    },
    "version": "2.1.0"
  },
  "paths": {
...
    "/downloads": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Downloads",
        "operationId": "get_downloads",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Download"
                  },
                  "type": "array",
                  "title": "Response Get Downloads"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "downloads"
        ],
        "summary": "Post Downloads",
        "operationId": "post_downloads",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Download"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/downloads/options/available": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Downloads Options Available",
        "operationId": "get_downloads_options_available",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableDownloadOptions"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/downloads/options/defaults": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Downloads Options Defaults",
        "operationId": "get_downloads_options_defaults",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadOptions"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/downloads/status": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Downloads Status",
        "operationId": "get_downloads_status",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/downloads/{download_id}": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Download",
        "operationId": "get_download",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Download"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "downloads"
        ],
        "summary": "Put Download",
        "operationId": "put_download",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Download"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "downloads"
        ],
        "summary": "Delete Download",
        "operationId": "delete_download",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/downloads/{download_id}/file": {
      "get": {
        "tags": [
          "downloads"
        ],
        "summary": "Get Download File",
        "operationId": "get_download_file",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "download_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Download Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "audio/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "video/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batch": {
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDownload"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "batch"
        ],
        "summary": "Put Batch",
        "operationId": "put_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDownloadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDownload"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "batch"
        ],
        "summary": "Post Batch",
        "operationId": "post_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDownloadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDownload"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "batch"
        ],
        "summary": "Delete Batch",
        "operationId": "delete_batch",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/batch/status": {
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "Get Batch Status",
        "operationId": "get_batch_status",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batch/file": {
      "get": {
        "tags": [
          "batch"
        ],
        "summary": "Get Batch File",
        "operationId": "get_batch_file",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioFormat": {
        "type": "string",
        "enum": [
          "aac",
          "flac",
          "m4a",
          "mp3",
          "opus",
          "wav"
        ],
        "title": "AudioFormat"
      },
      "AvailableDownloadOptions": {
        "properties": {
          "format": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AudioFormat"
                },
                {
                  "$ref": "#/components/schemas/VideoFormat"
                }
              ]
            },
            "type": "array",
            "title": "Format",
            "default": [
              "aac",
              "flac",
              "m4a",
              "mp3",
              "opus",
              "wav",
              "avi",
              "flv",
              "mkv",
              "mov",
              "mp4",
              "webm"
            ]
          },
          "quality": {
            "items": {
              "$ref": "#/components/schemas/DownloadQuality"
            },
            "type": "array",
            "title": "Quality",
            "default": [
              "best",
              "worst"
            ]
          },
          "embed_metadata": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "title": "Embed Metadata",
            "default": [
              true,
              false
            ]
          },
          "embed_thumbnail": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "title": "Embed Thumbnail",
            "default": [
              true,
              false
            ]
          },
          "embed_subtitles": {
            "items": {
              "type": "boolean"
            },
            "type": "array",
            "title": "Embed Subtitles",
            "default": [
              true,
              false
            ]
          },
          "preferred_subtitles_language": {
            "items": {
              "$ref": "#/components/schemas/LanguageCode"
            },
            "type": "array",
            "title": "Preferred Subtitles Language",
            "default": [
              "ar",
              "de",
              "en",
              "es",
              "fr",
              "hi",
              "it",
              "ja",
              "ko",
              "pt",
              "ru",
              "tr",
              "zh"
            ]
          }
        },
        "type": "object",
        "title": "AvailableDownloadOptions"
      },
      "BatchDownload": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BatchDownloadStatus"
          },
          "videos": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Video"
            },
            "type": "object",
            "title": "Videos"
          },
          "urls": {
            "items": {
              "type": "string",
              "maxLength": 2083,
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "maxItems": 25,
            "minItems": 1,
            "title": "Urls"
          },
          "options": {
            "$ref": "#/components/schemas/DownloadOptions"
          }
        },
        "type": "object",
        "required": [
          "status",
          "videos",
          "urls",
          "options"
        ],
        "title": "BatchDownload"
      },
      "BatchDownloadInput": {
        "properties": {
          "urls": {
            "items": {
              "type": "string",
              "maxLength": 2083,
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "maxItems": 25,
            "minItems": 1,
            "title": "Urls"
          },
          "options": {
            "$ref": "#/components/schemas/DownloadOptions"
          }
        },
        "type": "object",
        "required": [
          "urls",
          "options"
        ],
        "title": "BatchDownloadInput"
      },
      "BatchDownloadStatus": {
        "properties": {
          "overall": {
            "$ref": "#/components/schemas/DownloadStatus"
          },
          "items": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DownloadStatus"
            },
            "type": "object",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "overall",
          "items"
        ],
        "title": "BatchDownloadStatus"
      },
      "Channel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "Channel"
      },
      "Download": {
        "properties": {
          "video": {
            "$ref": "#/components/schemas/Video"
          },
          "options": {
            "$ref": "#/components/schemas/DownloadOptions"
          },
          "status": {
            "$ref": "#/components/schemas/DownloadStatus"
          }
        },
        "type": "object",
        "required": [
          "video",
          "options",
          "status"
        ],
        "title": "Download"
      },
      "DownloadInput": {
        "properties": {
          "video": {
            "$ref": "#/components/schemas/Video"
          },
          "options": {
            "$ref": "#/components/schemas/DownloadOptions"
          }
        },
        "type": "object",
        "required": [
          "video",
          "options"
        ],
        "title": "DownloadInput"
      },
      "DownloadOptions": {
        "properties": {
          "format": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudioFormat"
              },
              {
                "$ref": "#/components/schemas/VideoFormat"
              }
            ],
            "title": "Format",
            "default": "mp4"
          },
          "quality": {
            "$ref": "#/components/schemas/DownloadQuality",
            "default": "best"
          },
          "embed_metadata": {
            "type": "boolean",
            "title": "Embed Metadata",
            "default": true
          },
          "embed_thumbnail": {
            "type": "boolean",
            "title": "Embed Thumbnail",
            "default": false
          },
          "embed_subtitles": {
            "type": "boolean",
            "title": "Embed Subtitles",
            "default": false
          },
          "preferred_subtitles_language": {
            "$ref": "#/components/schemas/LanguageCode",
            "default": "en"
          }
        },
        "type": "object",
        "title": "DownloadOptions"
      },
      "DownloadQuality": {
        "type": "string",
        "enum": [
          "best",
          "worst"
        ],
        "title": "DownloadQuality"
      },
      "DownloadState": {
        "type": "string",
        "enum": [
          "WAITING",
          "DOWNLOADING",
          "PROCESSING",
          "DONE",
          "ERROR"
        ],
        "title": "DownloadState"
      },
      "DownloadStatus": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/DownloadState"
          },
          "downloaded_bytes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Downloaded Bytes"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          },
          "elapsed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed"
          },
          "eta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "postprocessor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postprocessor"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "state",
          "progress"
        ],
        "title": "DownloadStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LanguageCode": {
        "type": "string",
        "enum": [
          "ar",
          "de",
          "en",
          "es",
          "fr",
          "hi",
          "it",
          "ja",
          "ko",
          "pt",
          "ru",
          "tr",
          "zh"
        ],
        "title": "LanguageCode"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Video": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "duration": {
            "type": "string",
            "title": "Duration"
          },
          "thumbnail": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Thumbnail"
          },
          "channel": {
            "$ref": "#/components/schemas/Channel"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "duration",
          "thumbnail",
          "channel"
        ],
        "title": "Video"
      },
      "VideoFormat": {
        "type": "string",
        "enum": [
          "avi",
          "flv",
          "mkv",
          "mov",
          "mp4",
          "webm"
        ],
        "title": "VideoFormat"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}